minor refactor to the user repository
Data should be accessed via the repository and not directly from the data sources.
Changes have been made to implement that.
Since we have just one data source for now we could work with everything directly on the repository layer and move all the datasource implementations to the repository, to reduce redundancy.